requirements.txt
Django==1.11.3
requests==2.18.2
openpyxl==2.4.8
django-daterange-filter==1.3.0
XlsxWriter==0.9.8
